Added more per server edit buttons on bot panel. Added delete button on queue panel for admins. Added per server edit for automated task thresholds. Fixed a large number of potential crashes with reading values from mongodb for requests and guild settings.

+++ Toaster/GuildSettings.h
@@ -56,7 +56,7 @@ public:
 	std::chrono::seconds announcement_cooldown{ 0 };
 
 	// Open request cap
-	std::uint8_t requestLimitPerUser = 15;
+	std::uint32_t requestLimitPerUser = 15;
 
 	// Ping Rules
 	bool bPingOnNew			{ true };
